python/requirements.txt
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.31.0